+++ cdap-ui/cypress/integration/lab.spec.ts
@@ -19,43 +19,43 @@ describe('Lab ', () => {
   });
 
   it('should have cdap-common-experiment disabled by default', () => {
-    cy.visit('/cdap/ns/default/lab');
+    cy.visit('/cdap/lab');
     cy.get(`${dataCy('cdap-common-experiment-switch')} input`).should('have.value', 'false');
   });
 
   describe(' toggle experiment wrapper ', () => {
     it('should show default component when cdap-common-experiment is disabled', () => {
-      cy.visit('/cdap/ns/default/lab-experiment-test');
+      cy.visit('/cdap/lab-experiment-test');
       cy.get(dataCy('default-feature-toggle-selector')).should('have.text', 'This is default component for the toggle.');
     });
 
     it('show experimental component when cdap-common-experiment is enabled', () => {
-      cy.visit('/cdap/ns/default/lab');
+      cy.visit('/cdap/lab');
       cy.get(`${dataCy('cdap-common-experiment-switch')} input`).should('have.value', 'false');
       cy.get(`${dataCy('cdap-common-experiment-switch')}`).click();
       cy.get(`${dataCy('cdap-common-experiment-switch')} input`).should('have.value', 'true');
 
-      cy.visit('/cdap/ns/default/lab-experiment-test');
+      cy.visit('/cdap/lab-experiment-test');
       cy.get(dataCy('experimental-feature-toggle-selector')).should('have.text', 'This is experimental component for the toggle.');
 
-      cy.visit('/cdap/ns/default/lab');
+      cy.visit('/cdap/lab');
       cy.get(`${dataCy('cdap-common-experiment-switch')}`).click();
     });
   });
 
   describe(' experiment wrapper ', () => {
     it('should not show experimental component when cdap-common-experiment is disabled', () => {
-      cy.visit('/cdap/ns/default/lab-experiment-test');
+      cy.visit('/cdap/lab-experiment-test');
       cy.get(dataCy('experimental-feature-selector')).should('not.exist');
     });
 
     it('should show experimental component when cdap-common-experiment is enabled', () => {
-      cy.visit('/cdap/ns/default/lab');
+      cy.visit('/cdap/lab');
       cy.get(`${dataCy('cdap-common-experiment-switch')} input`).should('have.value', 'false');
       cy.get(`${dataCy('cdap-common-experiment-switch')}`).click();
       cy.get(`${dataCy('cdap-common-experiment-switch')} input`).should('have.value', 'true');
 
-      cy.visit('/cdap/ns/default/lab-experiment-test');
+      cy.visit('/cdap/lab-experiment-test');
       cy.get(dataCy('experimental-feature-selector')).should('have.text','This is an experimental component.');
     });
   });
